Add Condition to ConfigSettings
"Disabled" configs use their default value. Usually, configs with conditions will only be enabled when another config has a certain value (e.g. configs that modify the details of a specific feature only need to appear if that feature is enabled), or when another mod is loaded (e.g. configs that exist for compat purposes)

In future will cause configs to only appear in GUI when the condition passes